pyproject.toml
[tool.poetry]
name = "oop-coursework"
version = "0.1.0"
description = ""
authors = ["ArtemSermyagin <artem.sermyagin@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^6.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
